requirements.txt
ExifRead
Pillow
PyMuPDF
SQLAlchemy
beautifulsoup4
flask
flask-cors
numpy
pandas
python-docx
python-pptx
striprtf
openpyxl
keras
tensorflow
gunicorn
psycopg2
pillow-avif-plugin
pymysql  
cx_Oracle
cryptography